+++ app/src/main/java/ie/app/minimap/data/local/dao/NodeDao.kt
@@ -21,8 +21,8 @@ interface NodeDao {
     suspend fun delete(node: Node)
 
     @Query("SELECT * from nodes WHERE id = :id")
-    fun getItem(id: Int): Flow<Node>
+    fun getItem(id: Long): Flow<Node>
 
     @Query("SELECT * FROM nodes WHERE floor_id = :floorId ORDER BY id ASC")
-    fun getNodeByFloor(floorId :Int): Flow<List<Node>>
+    fun getNodeByFloor(floorId :Long): Flow<List<Node>>
 }
